Reorganizes packages and fixes package declaration ordering issue.
