+++ src/main/java/org/fcrepo/migration/validator/api/ReportGenerator.java
@@ -1,4 +1,4 @@
-package org.fcrepo.migration.validator;/*
+/*
  * Licensed to DuraSpace under one or more contributor license agreements.
  * See the NOTICE file distributed with this work for additional information
  * regarding copyright ownership.
@@ -15,11 +15,12 @@ package org.fcrepo.migration.validator;/*
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.fcrepo.migration.validator.api;
 
 /**
- * A runnable responsible for executing validations and processing the results.
- *
+ * This interface is responsible for transforming the on disk results of a validation into a
+ * a human readable report.
  * @author dbernstein
  */
-public abstract class ValidationTask implements Runnable {
+public interface ReportGenerator {
 }
