+++ src/main/java/org/fcrepo/migration/validator/api/ValidationConfig.java
@@ -1,4 +1,4 @@
-package org.fcrepo.migration.validator;/*
+/*
  * Licensed to DuraSpace under one or more contributor license agreements.
  * See the NOTICE file distributed with this work for additional information
  * regarding copyright ownership.
@@ -15,19 +15,37 @@ package org.fcrepo.migration.validator;/*
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.fcrepo.migration.validator.api;
 
-import java.util.List;
+import org.apache.commons.lang3.builder.ToStringBuilder;
+
+import static org.apache.commons.lang3.builder.ToStringStyle.MULTI_LINE_STYLE;
 
 /**
- * This class is responsible for writing validation results to disk.
+ * A data class for holding configuration information for a validation run
  *
  * @author dbernstein
  */
-public interface ValidationResultWriter {
+public abstract class ValidationConfig {
+
+    private int threadCount = 1;
+
+    @Override
+    public String toString() {
+        return ToStringBuilder.reflectionToString(this, MULTI_LINE_STYLE);
+    }
+
     /**
-     * Write the result to disk
-     *
-     * @param results The results to write
+     * @return
      */
-    void write(final List<ValidationResult> results);
+    public int getThreadCount() {
+        return this.threadCount;
+    }
+
+    /**
+     * @param threadCount
+     */
+    public void setThreadCount(final int threadCount) {
+        this.threadCount = threadCount;
+    }
 }
