+++ src/main/java/org/fcrepo/migration/validator/api/ValidationExecutionState.java
@@ -1,4 +1,3 @@
-package org.fcrepo.migration.validator;
 /*
  * Licensed to DuraSpace under one or more contributor license agreements.
  * See the NOTICE file distributed with this work for additional information
@@ -16,16 +15,12 @@ package org.fcrepo.migration.validator;
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.fcrepo.migration.validator.api;
 
 /**
- * Fedora 3 source types
+ * A data class that holds state information related to an executing validation run.A
  *
  * @author dbernstein
  */
-public enum F3SourceTypes {
-    AKUBRA, LEGACY, EXPORTED;
-
-    static F3SourceTypes toType(final String v) {
-        return valueOf(v.toUpperCase());
-    }
+public class ValidationExecutionState {
 }
