+++ src/main/java/org/fcrepo/migration/validator/api/ValidationResult.java
@@ -1,4 +1,4 @@
-package org.fcrepo.migration.validator;/*
+/*
  * Licensed to DuraSpace under one or more contributor license agreements.
  * See the NOTICE file distributed with this work for additional information
  * regarding copyright ownership.
@@ -15,11 +15,14 @@ package org.fcrepo.migration.validator;/*
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.fcrepo.migration.validator.api;
 
 /**
- * A data class that holds state information related to an executing validation run.A
+ * The result of a validation operation. A validation result corresponds to the most granular level of information
+ * communicated in the validation report.
  *
  * @author dbernstein
  */
-public class ValidationExecutionState {
+public interface ValidationResult {
+
 }
