+++ src/main/java/org/fcrepo/migration/validator/api/ValidationResultWriter.java
@@ -1,4 +1,4 @@
-package org.fcrepo.migration.validator;/*
+/*
  * Licensed to DuraSpace under one or more contributor license agreements.
  * See the NOTICE file distributed with this work for additional information
  * regarding copyright ownership.
@@ -15,10 +15,20 @@ package org.fcrepo.migration.validator;/*
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.fcrepo.migration.validator.api;
+
+import java.util.List;
 
 /**
- * This concrete implementation generates an HTML-based report.
+ * This class is responsible for writing validation results to disk.
+ *
  * @author dbernstein
  */
-public class HtmlReportGenerator implements ReportGenerator {
+public interface ValidationResultWriter {
+    /**
+     * Write the result to disk
+     *
+     * @param results The results to write
+     */
+    void write(final List<ValidationResult> results);
 }
