+++ src/main/java/org/fcrepo/migration/validator/impl/F3SourceTypes.java
@@ -1,4 +1,4 @@
-package org.fcrepo.migration.validator;/*
+/*
  * Licensed to DuraSpace under one or more contributor license agreements.
  * See the NOTICE file distributed with this work for additional information
  * regarding copyright ownership.
@@ -15,13 +15,17 @@ package org.fcrepo.migration.validator;/*
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.fcrepo.migration.validator.impl;
 
 /**
- * The result of a validation operation. A validation result corresponds to the most granular level of information
- * communicated in the validation report.
+ * Fedora 3 source types
  *
  * @author dbernstein
  */
-public interface ValidationResult {
+public enum F3SourceTypes {
+    AKUBRA, LEGACY, EXPORTED;
 
+    public static F3SourceTypes toType(final String v) {
+        return valueOf(v.toUpperCase());
+    }
 }
